+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/instructions/InstructionProcessor.java
@@ -148,6 +148,10 @@ public interface InstructionProcessor extends ContextlessInstructionCreator, Mes
     /// @return true if the built-in variable is volatile.
     boolean isVolatileBuiltIn(String builtin);
 
+    /// Mindustry Logic is unable to parse some specific integer literal values. This method identifies supported
+    /// and unsupported integer literals.
+    boolean isValidIntegerLiteral(long value);
+
     /// Rewrites the literal to conform to mlog limitations. If such a conversion isn't possible, an empty optional
     /// is returned.
     ///